libublk: move for_each_dev_id() and related tests into UblkCtrl
